requirements.txt
streamlit==1.28.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
plotly==5.18.0
folium==0.15.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.6.0
sqlalchemy==2.0.23
geopandas==0.14.1
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
newsapi-python==0.2.7
yfinance==0.2.32
streamlit-folium==0.15.0
aiohttp==3.9.1
asyncio==3.4.3
beautifulsoup4==4.12.2
lxml==4.9.3 